Extend Record.sub_tags() behavior.
Without positions arguments `sub_tags()` now returns all direct sub-tags
of a record. If positional arguments are provided they can be
hierarchical tags (slash-separated), and in that case all sub-tags that
match any of the provided tags are turned.

Also I did an initial run of `mypy` with relaxed rules and fixed few of
the potential issues reported by `mypy`.